AddActivity -> UpsertFragment 교체 및 이에 맞게 리팩토링
